// src/modules/invoice/usecase/find-usecase/find-invoice-usecase.dto.ts
export interface InputFindInvoiceDTO {
    id: string;
}

export interface OutputFindInvoiceDTO {
    id: string;
    name: string;
    document: string;
    address: {
        street: string;
        number: string;
        complement: string;
        city: string;
        state: string;
        zipCode: string;
    };
    items: {
        id: string;
        name: string;
        price: number;
    }[];
    total: number;
    createdAt: Date;
}